refactor: modularize ai_ml utilities
