+++ src/ai_ml/utils/__init__.py
@@ -1,0 +1,35 @@
+from .config import (
+    config_loader,
+    setup_environment_variables,
+    get_api_keys,
+    get_config_manager,
+)
+from .logging_utils import logger_setup
+from .performance import (
+    performance_monitor,
+    PerformanceDataStore,
+    JSONPerformanceDataStore,
+    set_performance_data_store,
+)
+from .environment import (
+    create_directory_structure,
+    validate_environment,
+    generate_environment_report,
+    cleanup_temp_files,
+)
+
+__all__ = [
+    "config_loader",
+    "setup_environment_variables",
+    "get_api_keys",
+    "get_config_manager",
+    "logger_setup",
+    "performance_monitor",
+    "PerformanceDataStore",
+    "JSONPerformanceDataStore",
+    "set_performance_data_store",
+    "create_directory_structure",
+    "validate_environment",
+    "generate_environment_report",
+    "cleanup_temp_files",
+]
